Allow encoded slashes (%2F) to be passed through by the apache proxy
Note: per the docs, taking the safer method of not decoding the %2Fs but jst passing them through (which is actually what we want behavior-wise).  It's the API service's responsibility to understand what to do with the %2F characters.

The current behavior that this is replacing is to throw a 404 when a user passes in a %2F in the URL